requirements.txt
discord
fuzzywuzzy
confusables
levenshtein
pyyaml
